refactor: simplify Studio implementation
